+++ crates/zizmor/src/audit/bot_conditions.rs
@@ -201,11 +201,9 @@ impl BotConditions {
         // always passes if the actor is dependabot[bot].
         match Self::walk_tree_for_bot_condition(&expr, true) {
             // We have a bot condition and it dominates the expression.
-            (Some(expr), true) => Some((Subfeature::from_spanned_expr(expr, 0), Confidence::High)),
+            (Some(expr), true) => Some((Subfeature::new(0, expr), Confidence::High)),
             // We have a bot condition but it doesn't dominate the expression.
-            (Some(expr), false) => {
-                Some((Subfeature::from_spanned_expr(expr, 0), Confidence::Medium))
-            }
+            (Some(expr), false) => Some((Subfeature::new(0, expr), Confidence::Medium)),
             // No bot condition.
             (..) => None,
         }